feat: add auto detect router loader
