+++ src/components/Router/getRoutes.js
@@ -1,12 +1,16 @@
 import { lazy } from 'react'
 import { flow, get } from 'lodash-es'
+import getRootPagesEntries from './getRootPagesEntries'
 
-const getRoutes = (pages, loaderMap = {}) => {
+const getRoutes = (pages, loaders, isRoot = false) => {
   const routes = flow(
-    () => Object.entries(pages),
+    () => {
+      return isRoot ? getRootPagesEntries(pages) : Object.entries(pages)
+    },
     (pagesEntries) => pagesEntries.reduce((collect, pagesEntry) => {
-      const [path, page] = pagesEntry
-      const fileName = path.match(/\.\/pages\/(.*)\.jsx$/)?.[1]
+      const [path, page, rootPath] = pagesEntry
+      const fileName = (isRoot ? rootPath : path).match(/\.\/pages\/(.*)\.jsx$/)?.[1]
+      const loaderPath = path.replace('index.jsx', 'index.loader.js')
       if (!fileName) {
         return collect
       }
@@ -16,10 +20,13 @@ const getRoutes = (pages, loaderMap = {}) => {
         .replace(/\/index/, '')
 
       const isIndex = fileName === 'index'
+      const pageLoader = get(loaders, loaderPath)
       collect.push({
         path: isIndex ? '/' : `/${normalizedPathName.toLowerCase()}/`,
         element: lazy(page),
-        loader: get(loaderMap, fileName, null)
+        loader: pageLoader
+          ? (...args) => pageLoader().then((module) => module.default(...args))
+          : null
       })
       return collect
     }, [])
