+++ src/components/Router/index.jsx
@@ -4,26 +4,12 @@ import {
   RouterProvider,
   defer
 } from 'react-router-dom'
-import { preload } from 'swr'
-import fetcher from '../../utils/fetcher'
-import getApiHost from '../../utils/getApiHost'
 import SkeletonHome from '../Skeleton/Home'
 import ErrorElement from './ErrorElement'
 import Layout from './Layout'
 import NavBar from '../NavBar'
+import getAuth from './getAuth'
 import 'react-loading-skeleton/dist/skeleton.css'
-
-const logoutUrl = import.meta.env.VITE_LOGOUT_URL
-const loginUrl = import.meta.env.VITE_LOGIN_URL
-const host = getApiHost('VITE_AWS_CHECK_AUTHORIZE')
-const awsHostPrefix = import.meta.env.VITE_AWS_HOST_PREFIX
-const authConfig = {
-  host,
-  url: `${awsHostPrefix}/checkAuthorize`,
-  options: {
-    errorMessage: '未登入'
-  }
-}
 
 const withErrorElement = (routes) => routes.map((item) => {
   const {
@@ -37,34 +23,27 @@ const withErrorElement = (routes) => routes.map((item) => {
 })
 
 const defaultAuth = { message: 'NO USER' }
-const getAuth = () => preload(authConfig, fetcher)
-  .then((res) => {
-    if (res.message === 'Unauthorized') {
-      throw new Error(res.message)
-    }
-    return res.message
-  })
-  .catch((e) => {
-    console.log(e)
-    window.location.href = (
-      window.IS_MOCK
-        ? window.location.href.replace(window.location.pathname, `${window.APP_BASENAME}/${loginUrl}`)
-        : logoutUrl
-    )
-    return 'ERROR'
-  })
+
 const Router = (props) => {
   const { routes, basename = '/', isAuthRoutes = true } = props
   const appBaseName = `${window.APP_BASENAME}${basename}`
   const totalRoutes = [
     {
       element: <Layout appBaseName={appBaseName} />,
-      loader: () => {
-        if (!isAuthRoutes) {
+      loader: async ({ request }) => {
+        const { pathname } = new URL(request.url)
+        const expectedAuthRoutes = [
+          'external'
+        ].some((authRoute) => pathname.startsWith(`/${authRoute}`))
+        if (!isAuthRoutes && !expectedAuthRoutes) {
           return defaultAuth
         }
 
-        return defer({ message: getAuth() })
+        const [error, auth, response] = await getAuth()
+        if (error) {
+          throw response
+        }
+        return defer({ message: auth })
       },
       children: withErrorElement([
         ...routes,
