+++ src/sites/external/pages/confirm/index.loader.js
@@ -1,0 +1,3 @@
+import loader from '../index.loader'
+
+export default loader
